test(oban): don't depend on localhost postgres to run tests
