+++ instrumentation/opentelemetry_oban/test/support/test_repo.ex
@@ -1,5 +1,10 @@
 defmodule TestRepo do
   use Ecto.Repo,
     otp_app: :opentelemetry_oban,
-    adapter: Ecto.Adapters.Postgres
+    adapter: Ecto.Adapters.SQLite3
+end
+
+defmodule PrepareOban do
+  use Ecto.Migration
+  def up, do: Oban.Migrations.up()
 end
